Add dedicated InputSidebar component, move ColumnSelectionPane within it
