+++ mathesar_ui/src/systems/data-explorer/input-sidebar/column-selection-pane/SelectableColumn.svelte
@@ -2,7 +2,7 @@
   import { createEventDispatcher } from 'svelte';
   import { Button } from '@mathesar-component-library';
   import ColumnName from '@mathesar/components/column/ColumnName.svelte';
-  import type { ColumnWithLink } from '../utils';
+  import type { ColumnWithLink } from '../../utils';
 
   const dispatch = createEventDispatcher();
