+++ mathesar_ui/src/systems/data-explorer/input-sidebar/column-selection-pane/TableGroupCollapsible.svelte
